chore(indexer): Add entrypoint for indexer
fix: linter
