+++ indexer/cmd/indexer-refresh/README.md
@@ -1,0 +1,4 @@
+# indexer/cmd/indexer-refresh
+
+Entrypoint for the new WIP indexer.   After project is deployed and stable we will delete the [indexer/main.go](../indexer/main.go) file and move [indexer-refresh/main.go](./main.go) in it's place.
+
